+++ src/main/java/de/jeff_media/JukeBoxPlus/Messages.java
@@ -6,6 +6,9 @@ public class Messages {
 
     final String ADDED_DISK,
             ALREADY_ADDED;
+
+    final String ENABLED, DISABLED, LOOP, SHUFFLE, AUTOSTART, STOP, RADIUS;
+
     final Main main;
 
     Messages(Main main) {
@@ -13,6 +16,13 @@ public class Messages {
 
         ADDED_DISK = getMsg(Config.MSG_ADDED_DISC,"Added {NAME} to this Jukebox");
         ALREADY_ADDED = getMsg(Config.MSG_ALREADY_ADDED,"This Jukebox already contains {NAME}.");
+        ENABLED = getMsg(Config.MSG_ENABLED,"&aEnabled");
+        DISABLED = getMsg(Config.MSG_DISABLED,"&cDisabled");
+        LOOP = getMsg(Config.MSG_LOOP, "&6Loop");
+        SHUFFLE = getMsg(Config.MSG_SHUFFLE,"&6Shuffle");
+        AUTOSTART = getMsg(Config.MSG_AUTOSTART, "&6Autostart");
+        STOP = getMsg(Config.MSG_STOP,"&6Stop");
+        RADIUS = getMsg(Config.MSG_RADIUS,"&6Radius");
     }
 
     private String getMsg(String path, String defaultText) {