+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-## 1.1.0
+## 1.0.3
+- Made GUI 100% translatable
+
+## 1.0.2 
+- Fixed loop mode being enabled after the looped record has already been removed.
+- Added automatic config updater
+- Added bStats
+
+## 1.0.1
 - Added automatic update checker